[refactor] Restructure and deduplicate test suite
- Rename test_heightmap → test_dem_processor, drop TestReadElevations
  (private method, covered by public interface tests)
- Extract TestNormalizeArray into new test_raster_utils
- Rename test_terrain_processing → test_sdf_model_writer
- Rename test_site_config → test_types; move load_sites_from_yaml
  tests into test_cli where the function lives
- Collapse redundant TestCatalogEntry/TestGetSite and
  TestSiteCatalog/TestListSites in test_site_catalog
- Fold TestIntegrationFullROIPipeline into TestIntegrationPipeline
- Merge duplicate cache-hit tests in test_downloader into one
