+++ lunar_terrain_exporter/CMakeLists.txt
@@ -25,12 +25,14 @@ if(BUILD_TESTING)
   ament_lint_auto_find_test_dependencies()
 
   find_package(ament_cmake_pytest REQUIRED)
-  ament_add_pytest_test(test_site_catalog test/test_site_catalog.py)
-  ament_add_pytest_test(test_site_config test/test_site_config.py)
   ament_add_pytest_test(test_cli test/test_cli.py)
+  ament_add_pytest_test(test_downloader test/test_downloader.py)
+  ament_add_pytest_test(test_dem_processor test/test_dem_processor.py)
+  ament_add_pytest_test(test_raster_utils test/test_raster_utils.py)
+  ament_add_pytest_test(test_site_catalog test/test_site_catalog.py)
+  ament_add_pytest_test(test_types test/test_types.py)
   ament_add_pytest_test(test_integration test/test_integration.py)
-  ament_add_pytest_test(test_heightmap test/test_heightmap.py)
-  ament_add_pytest_test(test_terrain_processing test/test_terrain_processing.py)
+  ament_add_pytest_test(test_sdf_model_writer test/test_sdf_model_writer.py)
 endif()
 
 ament_package()
