+++ lunar_terrain_exporter/test/test_downloader.py
@@ -23,30 +23,19 @@ from lunar_terrain_exporter.utils.file_downloader import FileDownloader
 
 
 class TestFileDownloader:
-    def test_cache_uses_url_filename(self):
-        """Cache path is simply cache_dir / filename from URL."""
+    def test_cache_hit_returns_file_without_downloading(self):
+        """If file exists in cache, return it by URL-derived filename without downloading."""
         with tempfile.TemporaryDirectory() as tmpdir:
             dl = FileDownloader(Path(tmpdir))
             url = 'https://pgda.gsfc.nasa.gov/data/LOLA_5mpp/Site01/Site01_final_adj_5mpp_surf.tif'
-            # Pre-create the file so download() returns the cached path
             expected = Path(tmpdir) / 'Site01_final_adj_5mpp_surf.tif'
             expected.write_bytes(b'data')
-            result = dl.download(url)
-            assert result == expected
-            assert result.name == 'Site01_final_adj_5mpp_surf.tif'
-
-    def test_returns_cached_file_without_download(self):
-        """If file exists in cache, return it without downloading."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            dl = FileDownloader(Path(tmpdir))
-            url = 'https://example.com/test.img'
-            cache_path = Path(tmpdir) / 'test.img'
-            cache_path.write_bytes(b'cached data')
 
             with patch('lunar_terrain_exporter.utils.file_downloader.requests') as mock_req:
                 result = dl.download(url)
                 mock_req.get.assert_not_called()
-            assert result == cache_path
+            assert result == expected
+            assert result.name == 'Site01_final_adj_5mpp_surf.tif'
 
     def test_downloads_when_not_cached(self):
         """Downloads and saves file when not in cache."""
