+++ lunar_terrain_exporter/test/test_types.py
@@ -13,17 +13,10 @@
 # limitations under the License.
 
 
-"""Tests for site configuration dataclasses and YAML loading."""
+"""Tests for site configuration dataclasses."""
 
-from pathlib import Path
-import tempfile
-
-from lunar_terrain_exporter.cli import load_sites_from_yaml
-from lunar_terrain_exporter.utils.types import (
-    BoundingBox, LunarSite, ROI,
-)
+from lunar_terrain_exporter.utils.types import BoundingBox, LunarSite, ROI
 import pytest
-import yaml
 
 
 class TestBoundingBox:
@@ -180,80 +173,6 @@ class TestLunarSite:
             ).validate()
 
 
-class TestLoadSites:
-    def _write_yaml(self, data: dict, path: Path) -> Path:
-        config_file = path / 'sites.yaml'
-        with open(config_file, 'w') as f:
-            yaml.dump(data, f)
-        return config_file
-
-    def test_load_single_site(self):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            config_file = self._write_yaml({
-                'sites': [{
-                    'site': 'connecting_ridge',
-                    'roi': {
-                        'use_full': False,
-                        'bounding_box': {'lat': -86.5, 'lon': -4.0},
-                    },
-                }]
-            }, Path(tmpdir))
-            sites = load_sites_from_yaml(config_file)
-            assert len(sites) == 1
-            assert sites[0].name == 'connecting_ridge'
-            assert sites[0].site_code == 'Site01'
-            assert sites[0].roi.bounding_box.width_km == 10.0
-
-    def test_load_multiple_sites(self):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            config_file = self._write_yaml({
-                'sites': [
-                    {
-                        'site': 'connecting_ridge',
-                        'roi': {
-                            'use_full': False,
-                            'bounding_box': {'lat': -86.0, 'lon': 0.0},
-                        },
-                    },
-                    {
-                        'site': 'shackleton_rim',
-                        'roi': {
-                            'use_full': False,
-                            'bounding_box': {
-                                'lat': -87.0, 'lon': 10.0,
-                                'width_km': 5.0, 'height_km': 5.0,
-                            },
-                        },
-                    },
-                ]
-            }, Path(tmpdir))
-            sites = load_sites_from_yaml(config_file)
-            assert len(sites) == 2
-            assert sites[1].roi.bounding_box.width_km == 5.0
-
-    def test_missing_required_field_raises(self):
-        """An entry without a 'site' key is skipped (warning printed)."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            config_file = self._write_yaml({
-                'sites': [{'name': 'bad'}]
-            }, Path(tmpdir))
-            sites = load_sites_from_yaml(config_file)
-            assert len(sites) == 0
-
-    def test_load_full_roi_site(self):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            config_file = self._write_yaml({
-                'sites': [{
-                    'site': 'connecting_ridge',
-                    'roi': {'use_full': True},
-                }]
-            }, Path(tmpdir))
-            sites = load_sites_from_yaml(config_file)
-            assert len(sites) == 1
-            assert sites[0].roi.use_full is True
-            assert sites[0].roi.bounding_box is None
-
-
 class TestFromCatalog:
     def test_creates_config_from_catalog_name(self):
         config = LunarSite.from_catalog('connecting_ridge')
@@ -280,59 +199,3 @@ class TestFromCatalog:
     def test_unknown_site_raises(self):
         with pytest.raises(KeyError, match='no_such_site'):
             LunarSite.from_catalog('no_such_site')
-
-
-class TestLoadSitesCatalogShorthand:
-    def _write_yaml(self, data: dict, path: Path) -> Path:
-        config_file = path / 'sites.yaml'
-        with open(config_file, 'w') as f:
-            yaml.dump(data, f)
-        return config_file
-
-    def test_catalog_shorthand(self):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            config_file = self._write_yaml({
-                'sites': [{
-                    'site': 'connecting_ridge',
-                    'roi': {'use_full': True},
-                }]
-            }, Path(tmpdir))
-            sites = load_sites_from_yaml(config_file)
-            assert len(sites) == 1
-            assert sites[0].name == 'connecting_ridge'
-            assert 'Site01' in sites[0].dem_url
-
-    def test_catalog_shorthand_with_bbox(self):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            config_file = self._write_yaml({
-                'sites': [{
-                    'site': 'shackleton_rim',
-                    'roi': {
-                        'use_full': False,
-                        'bounding_box': {'lat': -86.5, 'lon': -4.0, 'width_km': 5.0},
-                    },
-                }]
-            }, Path(tmpdir))
-            sites = load_sites_from_yaml(config_file)
-            assert sites[0].roi.bounding_box.width_km == 5.0
-            assert 'Site04' in sites[0].dem_url
-
-    def test_mixed_catalog_and_explicit(self):
-        """Two catalog entries in a batch."""
-        with tempfile.TemporaryDirectory() as tmpdir:
-            config_file = self._write_yaml({
-                'sites': [
-                    {
-                        'site': 'connecting_ridge',
-                        'roi': {'use_full': True},
-                    },
-                    {
-                        'site': 'shackleton_rim',
-                        'roi': {'use_full': True},
-                    },
-                ]
-            }, Path(tmpdir))
-            sites = load_sites_from_yaml(config_file)
-            assert len(sites) == 2
-            assert sites[0].name == 'connecting_ridge'
-            assert sites[1].name == 'shackleton_rim'
